move message handling methods to separate class
