+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/network/MessageHandler.java
@@ -1,0 +1,136 @@
+package nl.tudelft.cs4160.trustchain_android.network;
+
+import android.util.Log;
+
+import com.google.protobuf.ByteString;
+
+import java.io.IOException;
+import java.util.List;
+
+import nl.tudelft.cs4160.trustchain_android.block.TrustChainBlockHelper;
+import nl.tudelft.cs4160.trustchain_android.block.ValidationResult;
+import nl.tudelft.cs4160.trustchain_android.message.MessageProto;
+import nl.tudelft.cs4160.trustchain_android.network.peer.Peer;
+import nl.tudelft.cs4160.trustchain_android.network.peer.PeerHandler;
+import nl.tudelft.cs4160.trustchain_android.storage.database.TrustChainDBHelper;
+
+public class MessageHandler {
+    private TrustChainDBHelper dbHelper;
+    private Network network;
+    private PeerHandler peerHandler;
+    final String TAG = "MesssageHandler";
+
+    public MessageHandler(Network network, TrustChainDBHelper dbHelper, PeerHandler peerHandler) {
+        this.network = network;
+        this.dbHelper = dbHelper;
+        this.peerHandler = peerHandler;
+    }
+
+    /**
+     * Handle an introduction request. Send a puncture request to the included invitee.
+     *
+     * @param peer    the peer requesting the introduction
+     * @param request the message.
+     * @throws IOException
+     */
+    public void handleIntroductionRequest(Peer peer, MessageProto.IntroductionRequest request) throws IOException {
+        peer.setConnectionType((int) request.getConnectionType());
+        if (getPeerHandler().size() > 1) {
+            Peer invitee = getPeerHandler().getEligiblePeer(peer);
+            if (invitee != null) {
+                network.sendIntroductionResponse(peer, invitee);
+                network.sendPunctureRequest(invitee, peer);
+                Log.d("Network", "Introducing " + invitee.getAddress() + " to " + peer.getAddress());
+            }
+        } else {
+            Log.d("Network", "Peerlist too small, can't handle introduction request");
+            network.sendIntroductionResponse(peer, null);
+        }
+    }
+
+    /**
+     * Handle an introduction response. Parse incoming PEX peers.
+     *
+     * @param peer    the peer that sent this response.
+     * @param response the message.
+     */
+    public void handleIntroductionResponse(Peer peer, MessageProto.IntroductionResponse response) throws Exception {
+        peer.setConnectionType((int) response.getConnectionType());
+        List<ByteString> pex = response.getPexList();
+        for (ByteString pexPeer : pex) {
+            Peer p = Peer.deserialize(pexPeer.toByteArray());
+            Log.d(TAG, "From " + peer + " | found peer in pexList: " + p);
+
+            if (!getPeerHandler().hashId.equals(p.getPeerId())) {
+                getPeerHandler().getOrMakePeer(p.getPeerId(), p.getAddress());
+            }
+        }
+    }
+
+    /**
+     * Handle a puncture request. Sends a puncture to the puncture inboxItem included in the message.
+     *
+     * @param peer    the origin inboxItem.
+     * @param request the message.
+     * @throws IOException
+     */
+    public void handlePunctureRequest(Peer peer, MessageProto.PunctureRequest request) throws IOException {
+        Peer puncturePeer = null;
+        try {
+            puncturePeer = Peer.deserialize(request.getPuncturePeer().toByteArray());
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();
+        }
+        if (!getPeerHandler().peerExistsInList(puncturePeer)) {
+            network.sendPuncture(puncturePeer);
+        }
+    }
+
+    /**
+     * Handle a puncture. Does nothing because the only purpose of a puncture is to punch a hole in the NAT.
+     *
+     * @param peer    the origin inboxItem.
+     * @param puncture the message.
+     * @throws IOException
+     */
+    public void handlePuncture(Peer peer, MessageProto.Puncture puncture) throws IOException {
+    }
+
+    /**
+     * Handle the received (half) block.
+     * This block is placed in in the TrustChainDB, except if it is INVALID.
+     * @param peer the sending peer
+     * @param block the data send
+     * @throws IOException
+     */
+    public void handleReceivedBlock(Peer peer, MessageProto.TrustChainBlock block) {
+        try {
+            if (TrustChainBlockHelper.validate(block,dbHelper).getStatus() != ValidationResult.INVALID ) {
+                dbHelper.replaceInDB(block);
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    /**
+     * Handle crawl request
+     * @param peer the sending peer
+     * @param request the crawlRequest
+     * @throws IOException
+     */
+    public void handleCrawlRequest(Peer peer, MessageProto.CrawlRequest request) throws IOException {
+        //ToDo for future application sending the entire chain is a bit too much
+        for (MessageProto.TrustChainBlock block : dbHelper.getAllBlocks()) {
+            network.sendBlockMessage(peer, block);
+        }
+    }
+
+    /**
+     * Return the peer handler object.
+     * @return
+     */
+    public PeerHandler getPeerHandler() {
+        return peerHandler;
+    }
+}
